Replace checkFK with swapFKReferences
Instead of merely checking if the foreign key constraints are valid,
swap the field_name/field_value mapping so that the values of the
foreign keys are replaced with a pointer to the field in the parent
table, so that

{ /* ... */ fk_field_name: fk_field_value }

becomes

{ /* ... */ fk_field_name: key(parent_table, fk_field_name) }

This is allowed since the foreign key field has to have the same name as
the field it references. Aliases can be supported by including the
original field name (in the parent table) inside the foreign key
metadata at the child table.