Use new index key structure
Previously, index keys followed this structure:

  table/index/<ipk> -> fk
  table/index/<ipk>/field_name_a -> value(field_name_a)
  ...
  table/index/<ipk>/field_name_n -> value(field_name_n)

Where `ipk` is an autoincremented index key, and `fk` is the primary
key of the original row of the indexed values. This means that (a
minimum of) two reads are necessary to get an indexed value.

The new (non-unique) index key structure is:

  table/index/field_name_a/<value(field_name_a)>/<value(fk)> -> ⊥
  ...
  table/index/field_name_n/<value(field_name_n)>/<value(fk)> -> ⊥

By embedding the actual value of the foreign key and the field, we only
need a single read to get all the information we need from an index, at
the cost of extra key size and non-linear scans. This last point can be
solved by using the orderedKeySet to store them.