Add keys to kset on successful put
Because the previous versions of put/get allowed lists of keys, and the
current structure of a key is a list (bucklescript adt representation),
we wrap their representation in a map { key: key_structure }. That way
we can still treat lists of keys the same way as before.

Given that antidote keys can only be strings, we convert the keys just
before inserting to the database. The kset will still contain the
original keys. This could be mitigated by implementing the kset in
javascript, instead of using the current ml version.