Implement serialize / deserialize for kset
Currently, keys (and the kset) are implemented as OCaml variant types,
and compiled with bucklescript to Javascript. As of bucklescript version
0.1.5, the js representation for variant types is an array, plus some
properties (tags) to identity the specific variant (as can be seen here
https://github.com/bloomberg/bucklescript/blob/f970f58b0e0a53594065c7a9a2901f8e0c624624/site/docsource/Runtime-representation.adoc).

By default, the data format for encoding js terms to antidote is
messagepack, although other formats can be used. The issue here is that
almost all formats ignore properties in arrays, resulting in corrupted
data when inserting the kset into antidote and reading it back.

To account for this, we implement a custom serialization format for kset
keys (or ocaml variant types in general), that transforms an array to an
object literal, preserving all the properties in the process.

This should fix issues with kv.runT, which would read the kset at the
beggining of a transaction, pass it to the function, and write it back
just before commit.