Track incident foreign keys in the metadata
When adding a foreign key between two tables, we used to add information
about that on the referencing table, and nothing on the other.

Now we add information to both tables - this way the referenced table
can know if there are tables pointing to it, which helps in the event of
updating a field that is a foreign key to other table.