Change fk representation
Previously a foreign key was represented as a pointer to the actual
value in the parent table. This presents some problems regarding key
serialization and also extra latency for reads, that have to perform an
extra get per row.

The new representation stores the actual value of the foreign key, and
adds new metadata to the parent table, that is responsible for updating
children tables when inserting and updating.