Change select implementation
The old version of select would receive an array of keys, and will fetch
all subfields for each key, computing each subkey manually.

The new version uses the `batch` and `subkeys` methods from the
orderedKeySet to quickly compute the necessary keys to fetch. In
addition, the signature has changed to be able to support generalized
queries against arbitrary fields, although only primary keys are
implemented for the time being.