Implement strict subkey variant
Last version of kset fixed a "bug" where subkeys would return the
calling key. This was wrong as a key can't be a subkey of itself.

Given that some parts of the code depended on the key being included, we
offer the old behavior with the subkeys function, and offer the new
strictSubKeys function for parts of the code that may not want the
original key to be included in the returned list.