Implement index pruning
As (unique) index keys contain field values, it becomes necessary to
remove (or update) those keys if/when the actual value of the field it
indexes changes. The current implementation takes a row marked for
deletion or a row that is affected by an update, and removes all the
appropiate (unique) index keys from the ordered key set.