Implement new multi-way join
Allows to join more than two tables with multiple predicates over them.
The previous design already allowed this, but only when the three tables
had a single field in common. This allows queries like:

SELECT *
  FROM x, y, z
 WHERE x.refY = y.id
   AND y.refZ = z.id
   AND x.some_other_field = ...

With this, a new join predicate syntax is used, which presents a
api-breaking change. The old version required a `using` key with an
array of fields in the same order as the tables given to the query. The
new desing looks as follows (using the previous statement as guidance):

using: [ { x: 'refY', y: 'id}, { y: 'refZ', z: 'id' } ]

Thus, the previous sql query would look like:

antidoteSQL.select(_, '*', ['x', 'y', 'z'], {
  using: [ { x: 'refY', y: 'id}, { y: 'refZ', z: 'id' } ],
  x: { some_other_field: /* ... */ }
}